+++ src/extension.ts
@@ -10,9 +10,9 @@ export function activate(context: vscode.ExtensionContext) {
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-	let disposable = vscode.commands.registerCommand('peeker.show', () => {
-    open(context);
-	});
+  let disposable = vscode.commands.registerCommand('peeker.show', () => {
+    open(context, vscode.window.activeTextEditor);
+  });
 
 	context.subscriptions.push(disposable);
 }