+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 All notable changes to the "vscode-peeker" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## v0.2.0
+2021-02-20
 
-## [Unreleased]
+- improved behaviour when updating selection
+- added context menu command to look up current selection
+- add "Open in browser" button
+
+## v0.1.0
+2021-01-07
 
 - Initial release